+++ source/ArchivedMessages.Application/Configuration/ArchivedMessageConfiguration.cs
@@ -12,12 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.Application.SearchMessages;
 using Energinet.DataHub.EDI.ArchivedMessages.Infrastructure;
 using Energinet.DataHub.EDI.ArchivedMessages.Infrastructure.Configuration.DataAccess;
 using Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
 using Energinet.DataHub.EDI.Common.Configuration;
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Energinet.DataHub.EDI.ArchivedMessages.Application.Configuration;
@@ -27,7 +25,6 @@ public static class ArchivedMessageConfiguration
     public static void Configure(IServiceCollection services)
     {
         services.AddScopedSqlDbContext<ArchivedMessagesContext>();
-        services.AddTransient<IRequestHandler<GetMessagesQuery, MessageSearchResult>, GetMessageQueryHandler>();
         services.AddScoped<IArchivedMessageRepository, ArchivedMessageRepository>();
         services.AddScoped<IArchivedMessagesClient, ArchivedMessagesClient>();
     }
