+++ source/ArchivedMessages.Infrastructure/QueryInput.cs
@@ -12,18 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Generic;
-using Energinet.DataHub.EDI.Application.Configuration.Queries;
-using NodaTime;
+using Dapper;
 
-namespace Energinet.DataHub.EDI.Application.SearchMessages;
+namespace Energinet.DataHub.EDI.ArchivedMessages.Infrastructure;
 
-public sealed record GetMessagesQuery(
-    MessageCreationPeriod? CreationPeriod = null,
-    string? MessageId = null,
-    string? SenderNumber = null,
-    string? ReceiverNumber = null,
-    IReadOnlyList<string>? DocumentTypes = null,
-    IReadOnlyList<string>? BusinessReasons = null) : IQuery<MessageSearchResult>;
-
-public record MessageCreationPeriod(Instant DateToSearchFrom, Instant DateToSearchTo);
+public record QueryInput(string SqlStatement, DynamicParameters Parameters);
