+++ source/ArchivedMessages.Interfaces/IArchivedMessagesClient.cs
@@ -37,4 +37,12 @@ public interface IArchivedMessagesClient
     /// <param name="cancellationToken"></param>
     /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
     Task<Stream?> GetAsync(string messageId, CancellationToken cancellationToken);
+
+    /// <summary>
+    /// Search for messages in the database
+    /// </summary>
+    /// <param name="queryInput"></param>
+    /// <param name="cancellationToken"></param>
+    /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
+    Task<MessageSearchResult> SearchAsync(GetMessagesQuery queryInput, CancellationToken cancellationToken);
 }
